fix(parsing): overflow too large values when reading
This shouldn't be a problem normally except the situation is unhandled. If the
supplied byte is, well, *not* a byte, then either the data is persisted even
though the layout is ignored, or on >2^32-1 / >2^64-1 we take the corrupt
"byte" and manually overflow - which seems what Emacs prefers or throws
`overflow-error`.