fix(emacs): require 27.x due to missing arbitrary precision ints
This is problem mainly for double-word counter and other parts which by default
won't fit and need to be split into (at least) two variables of the patching
size + re-calc algo to either assemble into the bignum or split into
high-/low-precision parts and calculate for/with each separately.

Currently I don't have a use-case for Emacs <27.x though.

Ref: https://github.com/emacs-mirror/emacs/blob/a2d4034/etc/NEWS.27#L2692